separate rai from lib
resolves https://github.com/wydengyre/raiplayrss/issues/40
