+++ biome.json
@@ -10,7 +10,7 @@
 		}
 	},
 	"files": {
-		"include": ["cf/**/*.ts", "cf/**/*.json","lib/**/*.ts", "lib/**/*.json", "tools/**/*.ts"],
+		"include": ["rai/**/*.ts", "rai/**/*.json", "server/**/*.ts", "server/**/*.json", "cf/**/*.ts", "cf/**/*.json", "tools/**/*.ts"],
 		"ignore": ["cf/.wrangler"]
 	}
 }
