+++ cf/tsconfig.json
@@ -1,9 +1,9 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "../lib/tsconfig.json",
+  "extends": "../server/tsconfig.json",
   "compilerOptions": {
 		"types": ["@cloudflare/workers-types"]
   },
 	"include": ["*.ts"],
-	"exclude": ["*.test.ts"]
+	"exclude": ["*.test.ts", "declarations.d.ts"]
 }
