+++ cf/worker.test.ts
@@ -5,7 +5,7 @@ import { getPodcastFromFeed } from "@podverse/podcast-feed-parser";
 import { createServerAdapter } from "@whatwg-node/server";
 import { Router, RouterType, error, json } from "itty-router";
 import { UnstableDevWorker, unstable_dev } from "wrangler";
-import genresJson from "../lib/test/generi.json" with { type: "json" };
+import genresJson from "../rai/test/generi.json" with { type: "json" };
 import feedJson from "./test/lastoriaingiallo.json" with { type: "json" };
 import expectedJson from "./test/lastoriaingiallo.parsed.json" with {
 	type: "json",
