+++ package.json
@@ -8,25 +8,30 @@
     "check": "npm run lint && npm run typecheck",
     "ci": "npm run build-podcast && npm run ci-check && npm run test",
     "ci-check": "npm run ci-lint && npm run typecheck",
-    "ci-lint": "biome ci cf lib",
+    "ci-lint": "biome ci rai server cf tools",
     "deploy": "wrangler -c cf/wrangler.toml --env prod deploy",
     "deploy-preview": "wrangler -c cf/wrangler.toml deploy --dry-run --outdir '../dist'",
     "dev": "wrangler -c cf/wrangler.toml dev",
-    "lint": "biome check --apply cf lib tools",
-    "test": "npm run test-lib && npm run test-cf",
+    "lint": "biome check --apply rai server cf tools",
+    "test": "npm run test-rai && npm run test-server && npm run test-cf",
     "test-cf": "tsx --tsconfig ./cf/tsconfig.json --test 'cf/**/*.test.ts'",
 		"test-cf-only": "tsx --tsconfig ./cf/tsconfig.json --test --test-only 'cf/**/*.test.ts'",
-    "test-lib": "tsx --tsconfig ./lib/tsconfig.json --test 'lib/**/*.test.ts'",
-		"test-lib-only": "tsx --tsconfig ./lib/tsconfig.json --test --test-only 'lib/**/*.test.ts'",
-    "typecheck": "concurrently \"npm run typecheck-lib\" \"npm run typecheck-cf\" \"npm run typecheck-tools\"",
+    "test-server": "tsx --tsconfig ./server/tsconfig.json --test 'server/**/*.test.ts'",
+		"test-server-only": "tsx --tsconfig ./server/tsconfig.json --test --test-only 'server/**/*.test.ts'",
+		"test-rai": "tsx --tsconfig ./rai/tsconfig.json --test 'rai/**/*.test.ts'",
+		"test-rai-only": "tsx --tsconfig ./rai/tsconfig.json --test --test-only 'rai/**/*.test.ts'",
+    "typecheck": "concurrently \"npm run typecheck-rai\" \"npm run typecheck-server\" \"npm run typecheck-cf\" \"npm run typecheck-tools\"",
     "//": "the output from typecheck can be hard to read, so we run them serially to make it easier",
-    "typecheck-serial": "npm run typecheck-lib && npm run typecheck-cf && npm run typecheck-tools",
+    "typecheck-serial": "npm run typecheck-rai && npm run typecheck-server && npm run typecheck-cf && npm run typecheck-tools",
     "typecheck-cf": "npm run typecheck-cf-prod && npm run typecheck-cf-test",
     "typecheck-cf-prod": "tsc --project cf/tsconfig.json",
     "typecheck-cf-test": "tsc --project cf/tsconfig.test.json",
-    "typecheck-lib": "npm run typecheck-lib-prod && npm run typecheck-lib-test",
-    "typecheck-lib-prod": "tsc --project lib/tsconfig.json",
-    "typecheck-lib-test": "tsc --project lib/tsconfig.test.json",
+    "typecheck-server": "npm run typecheck-server-prod && npm run typecheck-server-test",
+    "typecheck-server-prod": "tsc --project server/tsconfig.json",
+    "typecheck-server-test": "tsc --project server/tsconfig.test.json",
+		"typecheck-rai": "npm run typecheck-rai-prod && npm run typecheck-rai-test",
+		"typecheck-rai-prod": "tsc --project rai/tsconfig.json",
+		"typecheck-rai-test": "tsc --project rai/tsconfig.test.json",
     "typecheck-tools": "tsc --project tools/tsconfig.json"
   },
   "author": "",
