+++ rai/declarations.d.ts
@@ -1,0 +1,1 @@
+declare module "@podverse/podcast-feed-parser";
