+++ server/feed-handler.test.ts
@@ -1,15 +1,15 @@
 import { strict as assert } from "node:assert";
 import test from "node:test";
 import { json } from "itty-router";
-import { feedHandler } from "./feed-handler.js";
-import { FetchWithErr, NotOk, OkResponse } from "./fetch.js";
-import * as logger from "./logger.js";
-import genresJson from "./test/generi.json";
-import feedJson from "./test/lastoriaingiallo.json";
-import expectedJson from "./test/lastoriaingiallo.parsed.json" with {
+import { FetchWithErr, NotOk, OkResponse } from "../rai/fetch.js";
+import genresJson from "../rai/test/generi.json";
+import feedJson from "../rai/test/lastoriaingiallo.json";
+import expectedJson from "../rai/test/lastoriaingiallo.parsed.json" with {
 	type: "json",
 };
-import { parseFeed } from "./test/parse-feed.js";
+import { parseFeed } from "../rai/test/parse-feed.js";
+import { feedHandler } from "./feed-handler.js";
+import * as logger from "./logger.js";
 
 test("feed-handler", async (t) => {
 	await t.test(rssFeedSuccess);
