+++ server/handler.ts
@@ -1,6 +1,6 @@
 import { Router, createResponse, error, html, text } from "itty-router";
+import { mkFetchWithErr } from "../rai/fetch.js";
 import { feedHandler } from "./feed-handler.js";
-import { mkFetchWithErr } from "./fetch.js";
 import { genresHtml } from "./genres.js";
 import { Logger } from "./logger.js";
