+++ server/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "https://json.schemastore.org/tsconfig",
+	"extends": "../rai/tsconfig.json",
+	"include": ["*.ts"],
+	"exclude": ["*.test.ts", "declarations.d.ts"]
+}
